implement step 4

+++ src/command.rs
@@ -8,7 +8,7 @@ pub enum Command {
 pub struct SetCommand {
     pub key: String,
     pub flags: u32,
-    pub expire_time: i32,
+    pub expire_time: i64,
     pub byte_count: usize,
     pub noreply: bool,
     pub data: Vec<u8>,
